Implemented Pie Chart on Dashboard screen. Also added vertical UIScrollView on Dashboard screen.
Added back and forwad button to change the month of transactions showing in the screen. Clicking on any of the button will increase or decrease month by 1 and will show the transactions of that particular month only.

Same functionality added on dashboard to show the pie chart of specific month.