Added new functionality of Budget.
Added new Menu in the slide bar for Budget.

Create budget, set limit of budget and add the transaction in that budget category.

On crossing budget above 80% of set limit, budget progress bar color will become yellow.

On crossing the budget limit, color will be set to red.

All the UIViews and their autolayout constraints being created by programm and not by the drag and drop.